requirements.txt
ntt-injector==1.1.0
ntt-json-model==1.0.2
ntt-observable-list==1.0.0
ntt-signal==1.0.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
